Combine all VirtualLists into virtual_list::VirtualList
